refactor: address pr comments
