+++ lib/actions/api.js
@@ -696,7 +696,7 @@ export function findStopsWithinBBox (params) {
 
 export const clearStops = createAction('CLEAR_STOPS_OVERLAY')
 
-// Realtime Vehcile positions query
+// Realtime Vehicle positions query
 
 const receivedVehiclePositions = createAction('REALTIME_VEHICLE_POSITIONS_RESPONSE')
 const receivedVehiclePositionsError = createAction('REALTIME_VEHICLE_POSITIONS_ERROR')
