+++ lib/actions/ui.js
@@ -70,11 +70,7 @@ export function matchContentToUrl (location) {
             strict: false
           })
           const patternId = subMatch?.params?.patternId
-          if (patternId) {
-            dispatch(setViewedRoute({ patternId, routeId: id }))
-          } else {
-            dispatch(setViewedRoute({ routeId: id }))
-          }
+          dispatch(setViewedRoute({ patternId, routeId: id }))
         } else {
           dispatch(setViewedRoute(null))
           dispatch(setMainPanelContent(MainPanelContent.ROUTE_VIEWER))
@@ -217,13 +213,7 @@ export function setViewedStop (payload) {
 
 const viewStop = createAction('SET_VIEWED_STOP')
 
-export function setHighlightedStop (payload) {
-  return function (dispatch, getState) {
-    dispatch(setHoveredStop(payload))
-  }
-}
-
-const setHoveredStop = createAction('SET_HOVERED_STOP')
+export const setHoveredStop = createAction('SET_HOVERED_STOP')
 
 export const setViewedTrip = createAction('SET_VIEWED_TRIP')
