+++ lib/components/map/connected-route-viewer-overlay.js
@@ -6,8 +6,10 @@ import { connect } from 'react-redux'
 const mapStateToProps = (state, ownProps) => {
   const viewedRoute = state.otp.ui.viewedRoute
 
-  const routeData = viewedRoute && state.otp.transitIndex.routes
-    ? state.otp.transitIndex.routes[viewedRoute.routeId] : null
+  const routeData =
+    viewedRoute && state.otp.transitIndex.routes
+      ? state.otp.transitIndex.routes[viewedRoute.routeId]
+      : null
   let filteredPatterns = routeData?.patterns
 
   // If a pattern is selected, hide all other patterns
