+++ lib/components/viewers/route-details.js
@@ -1,128 +1,22 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import styled from 'styled-components'
 import { Button } from 'react-bootstrap'
 import { connect } from 'react-redux'
 
 import Icon from '../narrative/icon'
 import { extractHeadsignFromPattern, getColorAndNameFromRoute } from '../../util/viewer'
 import { getVehiclePositionsForRoute, findStopsForPattern } from '../../actions/api'
-import { setHighlightedStop, setMainPanelContent, setViewedStop, setViewedRoute } from '../../actions/ui'
+import { setHoveredStop, setMainPanelContent, setViewedStop, setViewedRoute } from '../../actions/ui'
 
-/** Converts text color (either black or white) to rgb */
-const textHexToRgb = (color) => (color === '#ffffff' ? '255,255,255' : '0,0,0')
-
-const Container = styled.div`
-  background-color: ${(props) => props.backgroundColor};
-  color: ${(props) => props.textColor};
-  overflow-y: hidden;
-  height: 100%;
-`
-const RouteNameContainer = styled.div`
-  padding: 8px;
-  color: ${props => props.textColor};
-`
-const LogoLinkContainer = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-`
-const MoreDetailsLink = styled.a`
-    color: ${(props) => props.color};
-    background-color: rgba(${(props) => textHexToRgb(props.color)},0.1);
-    padding: 5px;
-    display: block;
-    border-radius: 5px;
-    transition: 0.1s all ease-in-out;
-
-    &:hover {
-        background-color: rgba(255,255,255,0.8);
-        color: #333;
-    }
-}
-`
-const PatternContainer = styled.div`
-  background-color: ${(props) => props.routeColor};
-  color: ${(props) => props.textColor};
-  display: flex;
-  justify-content: flex-start;
-  align-items: baseline;
-  gap: 16px;
-  padding: 0 8px;
-  margin: 0;
-
-  overflow-x: scroll;
-
-  h4 {
-      margin-bottom: 0px;
-      white-space: nowrap;
-  }
-
-  button {
-      color: inherit;
-      border-bottom: 3px solid ${(props) => props.textColor};
-  }
-
-  button:hover, button:focus, button:visited {
-    text-decoration: none;
-  }
-  button:hover,  button:focus, button.active {
-      color: ${(props) => props.color};
-      background-color: ${(props) => props.textColor};
-  }
-}
-`
-
-const StopContainer = styled.div`
-  color: #333;
-  background-color: #fff;
-  overflow-y: scroll;
-  height: 100%;
-  /* 150px bottom padding is needed to ensure all stops 
-  are shown when browsers don't calculate 100% sensibly */
-  padding: 15px 0 100px;
-`
-const Stop = styled.a`
-  cursor: pointer;
-  color: #333;
-  display: block;
-  white-space: nowrap;
-  margin-left: 45px;
-  /* negative margin accounts for the height of the stop blob */
-  margin-top: -25px;
-
-  &:hover {
-    color: #23527c;
-  }
-
-  /* this is the station blob */
-  &::before {
-    content: '';
-    display: block;
-    height: 20px;
-    width: 20px;
-    border: 5px solid ${props => props.routeColor};
-    background: #fff;
-    position: relative;
-    top: 20px;
-    left: -35px;
-    border-radius: 20px;
-  }
-
-  /* this is the line between the blobs */
-  &::after {
-    content: '';
-    display: block;
-    height: 20px;
-    width: 10px;
-    background: ${props => props.routeColor};
-    position: relative;
-    left: -30px;
-    /* this is 2px into the blob (to make it look attached) + 30px so that each
-    stop's bar connects the previous bar with the current one */
-    top: -37px;
-  }
-`
+import {
+  Container,
+  RouteNameContainer,
+  LogoLinkContainer,
+  MoreDetailsLink,
+  PatternContainer,
+  StopContainer,
+  Stop
+} from './styled'
 
 class RouteDetails extends Component {
   static propTypes = {
@@ -169,7 +63,7 @@ class RouteDetails extends Component {
   }
 
   render () {
-    const { className, operator, pattern, route, setHighlightedStop } = this.props
+    const { className, operator, pattern, route, setHoveredStop } = this.props
     const { agency, patterns, url } = route
 
     const { backgroundColor: routeColor, color: textColor } = getColorAndNameFromRoute(
@@ -241,15 +135,15 @@ class RouteDetails extends Component {
         </PatternContainer>
         {pattern && (
           <StopContainer
-            onMouseLeave={() => setHighlightedStop(null)}
+            onMouseLeave={() => setHoveredStop(null)}
             routeColor={routeColor}
           >
             {pattern?.stops?.map((stop) => (
               <Stop
                 key={stop.id}
                 onClick={() => this._stopLinkClicked(stop.id)}
-                onFocus={() => setHighlightedStop(stop.id)}
-                onMouseOver={() => setHighlightedStop(stop.id)}
+                onFocus={() => setHoveredStop(stop.id)}
+                onMouseOver={() => setHoveredStop(stop.id)}
                 routeColor={routeColor.includes('ffffff') ? '#333' : routeColor}
               >
                 {stop.name}
@@ -272,7 +166,7 @@ const mapStateToProps = (state, ownProps) => {
 const mapDispatchToProps = {
   findStopsForPattern,
   getVehiclePositionsForRoute,
-  setHighlightedStop,
+  setHoveredStop,
   setMainPanelContent,
   setViewedRoute,
   setViewedStop
